@@ -31,13 +31,14 @@ classifiers =
 packages = find:
 zip_safe = True # if using mypy, must be False
 include_package_data = True
-python_requires = >=3.6
+python_requires = >=3.7
 install_requires =
     setuptools >= 38.4
     # additional packages
     typing-extensions
     numpy
     scipy
+    custom-inherit
 
 
 setup_requires =
