@@ -11,10 +11,15 @@ from .norofrenkel import NoroFrenkelPair
 from .potentials import CubicTable, Phi_hs, Phi_lj, Phi_nm, Phi_sw, Phi_yk, factory_phi
 from .potentials_base import PhiAnalytic, PhiBase, PhiCut, PhiGeneric, PhiLFS
 
+# updated versioning scheme
 try:
-    import pkg_resources
+    from importlib.metadata import version as _version
+except ImportError:
+    # if the fallback library is missing, we are doomed.
+    from importlib_metadata import version as _version  # type: ignore[no-redef]
 
-    __version__ = pkg_resources.get_distribution("MATS").version
+try:
+    __version__ = _version("analphipy")
 except Exception:
     # Local copy or not installed with setuptools.
     # Disable minimum version checks on downstream libraries.
